* It's now possible to set parameters in NodeManager through configuration data object * Managers print their configuration to trace.log when they change from the default * Manager::onConfig() now gives a pointer to the relevant metadata for the manager rather than the full data object. * Made sure Release(WIDCOMM) configuration compiles properly for Haggle and LuckyMe

+++ src/hagglekernel/Node.h
@@ -282,11 +282,11 @@ public:
 
         DataObjectRef getDataObject(bool withBloomfilter = true) const;
 		
-	unsigned int getMatchingThreshold() const { return matchThreshold; }
-	unsigned int getMaxDataObjectsInMatch() const { return numberOfDataObjectsPerMatch; }
+	unsigned long getMatchingThreshold() const { return matchThreshold; }
+	unsigned long getMaxDataObjectsInMatch() const { return numberOfDataObjectsPerMatch; }
 
-	void setMatchingThreshold(unsigned int value) { matchThreshold = value; }
-	void setMaxDataObjectsInMatch(unsigned int value) { numberOfDataObjectsPerMatch = value; }
+	void setMatchingThreshold(unsigned long value) { matchThreshold = value; }
+	void setMaxDataObjectsInMatch(unsigned long value) { numberOfDataObjectsPerMatch = value; }
 
         // Wrappers for adding, removing and updating attributes in
         // the node description associated with this node
